django-hackathon-starter/requirements.txt
Babel==1.3
Django==1.7.6
Jinja2==2.7.3
MarkupSafe==0.23
PyTumblr==0.0.6
Pygments==2.0.2
Sphinx==1.3.1
alabaster==0.7.3
astroid==1.3.6
beautifulsoup4==4.3.2
coverage==4.0a5
django-bootstrap-form==3.2
django-bootstrap-forms==0.1
django-bower==5.0.2
django-cors-headers==1.0.0
django-nose==1.3
djangorestframework==3.1.1
docutils==0.12
httplib2==0.9
httpretty==0.8.8
logilab-common==0.63.2
mock==1.0.1
nose==1.3.4
pylint==1.4.3
python-openid==2.2.5
pytz==2015.2
requests==2.6.0
requests-oauthlib==0.4.2
simplejson==3.6.5
six==1.9.0
snowballstemmer==1.2.0
sphinx-rtd-theme==0.1.7
twilio==3.8.0
googlemaps==2.2
